Move callback() invocation into request.get()
I was working with the Tracks Logins in MixPanel rule
on the website and the callback() invocation was
throwing three errors (callback not defined,
user not defined, context not defined), and upon
invocation was failing with:
 ERROR: Invalid response code from the auth0-sandbox: HTTP 400.
 Unable to compile submitted JavaScript.
 SyntaxError: Unexpected identifier SyntaxError: Unexpected identifier
I modified the script as in this request and it worked.

This change, if valid, will probably need to be applied to
other places as well, like the website.