Fix bug with properly supporting a state value
If Auth0 passes a state value to the webtask, the webtask needs to echo it back.The previous code wasn't doing this properly with the webtask's form method being set to 'get'. A 'post' sends the state param via query and 'confirm' value via post. The rule needs to be modified to then look for 'confirm' in the body vs. query.