Fix primaryUser issue
If the primary account changes, previously it would cause an error (in the code flow) or an id token with the wrong sub claim (user_id). We added a mechanism to indicate the new primary user ID in a rule to fix this issue.

Customers can set the primary user ID in a rule using `context.primaryUser = 'auth0|foo123'` after account linking.