Switch from using POST to GET to work around 'context' issues
Neither 'context.body' nor 'context.data' are filled using the written
documentation located here:
https://auth0.com/docs/multifactor-authentication/yubikey and the
Webtask setup instructions located here:
https://manage.auth0.com/#/tenant/webtasks

As a result, the POST variables (such as 'otp') and secrets (such as
'yubikey_clientid') aren't available, resulting in errors communicating
to YubiCo's API.

My hunch is that our token is missing 'pb' (see:
https://webtask.io/docs/token) but using 'wt-cli', I can't seem to
customize my token and there appears to be no way in the Auth0 dashboard
to do so, either.

I don't like using GET for this, it's not correct, but as a
proof-of-concept for how this is broken and how it could be worked
around, I'm adding it for Auth0 to take a look.