Delete custom-scopes.md
`context.jwtConfiguration` was used in the old non-OIDC conformant flows. The option we have now is setting `context.accessToken.scope`, but this behaves differently from what's mentioned in this rule.

We should either delete the current rule (it appears in search results, creates confusion) or update it. This PR is to delete the rule until we might add a better OIDC-conformant one later.